problem three unsolved issue
problem three unsolved issue , Algorithm too slow , need better solution

+++ euler/SDViewController.h
@@ -22,5 +22,5 @@
 - (NSMutableArray *) arrayofMultiples;
 - (int) sumofMultiples;
 - (int) problemtwo;
-- (int) problemthree;
+- (long long) problemthree;
 @end
